Add Minus Operation for TAction Finish Scripts

+++ src/Ext/Trigger/Hooks.cpp
@@ -9,7 +9,7 @@ DEFINE_HOOK(0x727064, TriggerTypeClass_HasLocalSetOrClearedEvent, 0x5)
 	GET(const int, nIndex, EDX);
 
 	return 
-		nIndex >= (int)PhobosTriggerEvent::LocalVariableGreaterThan && nIndex <= (int)PhobosTriggerEvent::LocalVariableAndIsTrue ?
+		nIndex >= PhobosTriggerEvent::LocalVariableGreaterThan && nIndex <= PhobosTriggerEvent::LocalVariableAndIsTrue ?
 		0x72706E : 
 		0x72707E;
 }
@@ -19,7 +19,7 @@ DEFINE_HOOK(0x727024, TriggerTypeClass_HasGlobalSetOrClearedEvent, 0x5)
 	GET(const int, nIndex, EDX);
 
 	return 
-		nIndex >= (int)PhobosTriggerEvent::GlobalVariableGreaterThan && nIndex <= (int)PhobosTriggerEvent::GlobalVariableAndIsTrue ?
+		nIndex >= PhobosTriggerEvent::GlobalVariableGreaterThan && nIndex <= PhobosTriggerEvent::GlobalVariableAndIsTrue ?
 		0x72702E : 
 		0x72703E;
 }
